Created ListAdapter
Created an adapter to work with the List. It extneds RecyclerView. It
has a function that will bind the view to the current position, as well
as function that returns the length of the array of the names in the
ImageList class.
It has an internal class that has a textview and imageview that will get
set based on which position you are in the list.