updated the ListFragment and added content
updated the ListFragment class with a RecyclerView object named
recyclerView and a ListAdapter object called listAdapter, creative I
know. The recyclerView sets the listAdapter as it's adapter with the
setAdapter method, then I created a LinearLayoutManager object called
layoutManager and set it as the LayoutManager for the recyclerView.

Also I added the animal photos to the drawable folder. Finished creating
the String[] with the animal names and the int[] with the corresponding
animnal photos from drawable.

Oh I also changed the theme to dark.