added comments and fixed fragment issue
every time the screen was rotated there was a new fragment created. So I
created a ListFragment object called savedFragment. There is then an if
statment that checks if the savedFragment is null. If it is then it
creates the fragment. Otherwise if it is not null then the fragment
already exists and doesn't need to be created again.