added new fragment to onAnimalSelected
So now when onAnimalSelected method runs it creates a new
ViewImageFragment. This fragment is then replaces the last fragment and
is added to the backstack so that we can go back. Fixed issue with
having the first fragment be on the backstack as well since when you
pressed back it would go to a blank screen.