Changed way fragments are found + other stuff
There was an issue when screen was rotated while inside
ImageViewFragment so had to create tags for the fragments and use
findViewByTag instead of findViewById to get it to work correcty.

Also created a bundle to pass the index of the selected animal to the
ViewImageFragment because it didn't want a custom constructor that
accepts the index. So that fragment now takes that index from the bundle
and uses it to change the Action bar to say the name of the selected
animal. Also did an Override on the onStop method to switch the Action
bar text back to the app_name from the strings.xml