created 2 new values files for different screen sizes
the new values files have a boolean value called "is_tablet" and it is
true if the device has a width of 600dp.

The main activity was also changed to reflect this. A boolean varibale
was created to keep track of whether the screen size is tablet or not.
An if statement was then created to run if it is not a tablet. The code
in that statement is the same code as before. The else statement has the
same code again but ListFragment is changed to GridFragment for tablet
view.