A lot of changes
Created GridFragment for tablet sized views. It was pretty similar to
the ListFragment so most of the code carried over.

Created a new Abstract class called Recycler adapter since both my
ListFragment and GridFragment would be using a similar adapter. Just
trying to minize repeated code. Created a GridAdapter that extends the
RecyclerAdapter. Also changed the ListAdapter to extend the new
RecyclerAdapter.

Renamed the fragment_listview.xml to fragment_recyclerview.xml since it
is now used for multiple things.

Also manged to get the ViewImageFragment to actually show an image. This
took hours and many failed attempts but finally found a way that works
well and of course it was way simpler than all the other methods I was
trying.

So close to being done this project.